code fix for placeholders and refactoring
